chore: added user login, jwt, usernames, updated tests

+++ backend/src/services/posts.js
@@ -1,6 +1,7 @@
 import { Post } from '../db/models/post.js'
-export async function createPost({ title, author, content, imageUrl, tags }) {
-  const post = new Post({ title, author, content, imageUrl, tags })
+import { User } from '../db/models/user.js'
+export async function createPost(userId, { title, content, imageUrl, tags }) {
+  const post = new Post({ title, author: userId, content, imageUrl, tags })
   return await post.save()
 }
 
@@ -13,8 +14,11 @@ async function listPosts(
 export async function listAllPosts(options) {
   return await listPosts({}, options)
 }
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options)
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername })
+  if (!user) return []
+
+  return await listPosts({ author: user._id }, options)
 }
 export async function listPostsByTag(tags, options) {
   return await listPosts({ tags }, options)
@@ -23,15 +27,16 @@ export async function getPostById(postId) {
   return await Post.findById(postId)
 }
 export async function updatePost(
+  userId,
   postId,
-  { title, author, content, imageUrl, tags },
+  { title, content, imageUrl, tags },
 ) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, content, imageUrl, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, content, imageUrl, tags } },
     { new: true },
   )
 }
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
